Removed timer from player class
