+++ DrawbackChess/Components/Pages/GameMenu.razor
@@ -18,6 +18,7 @@
     <div class="form_field">
         <label>Select Play Time (minutes):</label>
         <select @bind="playTime">
+            <option value="1">1</option>
             <option value="5">5</option>
             <option value="10">10</option>
             <option value="15">15</option>
